Compilable And Runable
can now compile and run the project to its former glory.

only issues is needing to setup a final running directory and move the
assets over so that it can load the python scripts and not error on not
being able to run the determined python script. :D. yay fucking works
finnaly. time for bed phew...