bpMesh - FIXES
added safety calls to bind to null objects so nothing else could modify
its contents. and fixed the draw direciton of one of the triangles as it
was wrong. (most of this class needs to be moves to a child class. and
this needs to be more or less converted into a base class to derive
from.